Capitalization conventions Simplify motor direction state machine- Encapsulate timer count/expired/enabled variables into a Counter struct Lots of reorganization Make project work better with CLion IDE

+++ Power_Board/include/stdhdr.h
@@ -31,7 +31,7 @@
 void block_ms(uint16_t ms);
 
 /** return the nearest integer within the given range */
-int clamp(int value, int lo, int hi);
+int16_t clamp(int16_t value, int16_t lo, int16_t hi);
 /** return the nearest float within the given range */
 float clamp_f(float value, float lo, float hi);
